src/data_preprocessing/ADV/requirements.txt
dill==0.4.0
numpy==1.26.4
scikit-learn==1.6.1
six==1.17.0
torchtext==0.9.0
tensorboard==2.20.0